requirements.txt
django
python-dotenv
djangorestframework
requests
Pillow
coreapi